Fix issue when combining other builders (build_cli, json_serializer) with realm_generator
