+++ generator/pubspec.yaml
@@ -12,7 +12,7 @@ environment:
   sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  analyzer: ^2.3.0
+  analyzer: ^3.0.0
   build_resolvers: ^2.0.0
   build: ^2.0.0
   dart_style: ^2.2.0
