+++ lib/src/cli/metrics/metrics_command.dart
@@ -15,22 +15,20 @@
 // limitations under the License.
 //
 ////////////////////////////////////////////////////////////////////////////////
-
 import 'dart:async';
+import 'dart:convert';
+import 'dart:io';
 
 import 'package:args/command_runner.dart';
+import 'package:crypto/crypto.dart';
+import 'package:logging/logging.dart';
+import 'package:path/path.dart' as path;
+import 'package:pub_semver/pub_semver.dart';
+import 'package:pubspec_parse/pubspec_parse.dart';
 
 import 'flutter_info.dart';
 import 'metrics.dart';
 import 'options.dart';
-
-import 'dart:convert';
-import 'dart:io';
-
-import 'package:crypto/crypto.dart';
-import 'package:logging/logging.dart';
-import 'package:pubspec_parse/pubspec_parse.dart';
-
 import 'utils.dart';
 
 class MetricsCommand extends Command<void> {
@@ -67,7 +65,7 @@ Future<void> uploadMetrics(Options options) async {
     return;
   }
 
-  final flutterInfo = await FlutterInfo.get(options);
+  final flutterInfo = await getInfo(options);
   final hostId = await machineId();
 
   final metrics = await generateMetrics(
@@ -163,3 +161,51 @@ extension _StringEx on String {
   static const _defaultSalt = <int>[75, 97, 115, 112, 101, 114, 32, 119, 97, 115, 32, 104, 101, 114];
   Digest strongHash({List<int> salt = _defaultSalt}) => sha256.convert([...salt, ...utf8.encode(this)]);
 }
+
+
+  Future<FlutterInfo?> getInfo(Options options) async {
+    final pubspecPath = options.pubspecPath;
+    final pubspec = Pubspec.parse(await File(pubspecPath).readAsString());
+
+    const flutter = 'flutter';
+    final flutterDep = pubspec.dependencies.values.whereType<SdkDependency>().where((dep) => dep.sdk == flutter).firstOrNull;
+    if (flutterDep == null) {
+      return null; // no flutter dependency, so not a flutter project
+    }
+
+    // Read constraints, if any
+    var flutterVersionConstraints = flutterDep.version.intersect(pubspec.environment?[flutter] ?? VersionConstraint.any);
+
+    // Try to read actual version from version file in .dart_tools. 
+    // This is updated when calling a flutter command on the project, 
+    // but not when calling a dart command..
+    final version = await safe(() async {
+      return Version.parse(await File(path.join(path.dirname(pubspecPath), '.dart_tool/version')).readAsString());
+    });
+
+    // Try to get full info by calling flutter executable
+    final info = await safe(() async {
+      final flutterRoot = options.flutterRoot;
+      final flutterPath = flutterRoot == null ? flutter : path.join(flutterRoot, 'bin', flutter);
+      final process = await Process.start(
+        flutterPath,
+        ['--version', '--machine'],
+      );
+      final infoJson = await process.stdout.transform(utf8.decoder).join();
+      return FlutterInfo.fromJson(json.decode(infoJson) as Map<String, dynamic>);
+    });
+
+    // Sanity check full info, if we have it
+    if (info != null && (version == null || version == info.frameworkVersion) && flutterVersionConstraints.allows(info.frameworkVersion)) {
+      // The returned info match both the projects constraints and the 
+      // flutter version of the lastest flutter command run on the project
+      return info; 
+    }
+
+    // Fallback to simplified info build from the version read from .dart_tool/version, 
+    // secondly the min constraint of the flutter SDK used
+    return FlutterInfo(
+      frameworkVersion: version ?? (await safe(() => (flutterVersionConstraints as VersionRange).min!)) ?? Version.none,
+      dartSdkVersion: Version.parse(Platform.version.toString().takeUntil(' ')),
+    );
+  }
